requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.0
Django==4.2.6
django-modeladmin-reorder==0.3.1
idna==3.4
pgeocode==0.4.1
psycopg2-binary==2.9.9
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.6
django-admin-interface==0.28.6
django-import==0.0.7
django-import-export==3.3.7
